% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_func_file.R
\name{DR_umap}
\alias{DR_umap}
\title{This function is used for dimension reduction for both training and test datasets, if dt.list.test is null
only output training dataset.}
\usage{
DR_umap(dt.list.train, dt.list.test = NULL)
}
\arguments{
\item{dt.list.train}{a list containing the multi-modality training features.}

\item{dt.list.test}{a list contraining the muli-modality test features or NULL.}
}
\value{
a list containing the training (and test) data matrix after dimension.
}
\description{
This function is used for dimension reduction for both training and test datasets, if dt.list.test is null
only output training dataset.
}
