% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_func_file.R
\name{draw_Heatmap}
\alias{draw_Heatmap}
\title{Draw the Heatmap to display the feature distribution.}
\usage{
draw_Heatmap(clust.data, out.clust, fig.name = "Heatmap")
}
\arguments{
\item{clust.data}{The datalist containing data.list.train, dt.list.test}

\item{out.clust}{The prediction cluster results}

\item{fig.name}{The name of heatmap}
}
\description{
Draw the Heatmap to display the feature distribution.
}
