% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintModels.r
\name{FormatCoefficientOutput.Text}
\alias{FormatCoefficientOutput.Text}
\title{FormatCoefficientOutput.Text}
\usage{
FormatCoefficientOutput.Text(
  Coefficient = NULL,
  SE = NULL,
  PValue = NULL,
  SpecialText = NULL,
  PrintStat = "SE",
  Digits = 4,
  IsPercent = F,
  CoefficientOnly = F,
  SpecialTextList = list(`-3` = "", `-2` = "no", `-1` = "yes")
)
}
\value{

}
\description{
FormatCoefficientOutput.Text
}
\keyword{internal}
