% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintModels.r
\name{GetPValueFromNormal}
\alias{GetPValueFromNormal}
\title{Get P-value from normal}
\usage{
GetPValueFromNormal(Stats)
}
\arguments{
\item{Stats}{}
}
\value{

}
\description{
Get P-value from normal
}
\keyword{internal}
