% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintModels.r
\name{GetStars.Text}
\alias{GetStars.Text}
\title{Get stars text}
\usage{
GetStars.Text(PValue, StarBreaks = c(-Inf, 0.01, 0.05, 0.1, Inf))
}
\arguments{
\item{StarBreaks}{}
}
\value{

}
\description{
Get stars text
}
\keyword{internal}
