% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_func_file.R
\name{knn_train_test}
\alias{knn_train_test}
\title{Title}
\usage{
knn_train_test(cmoic.clust, clust.data.train, clust.data.test)
}
\arguments{
\item{cmoic.clust}{The consensus clustering result variable}

\item{clust.data.train}{The training list of matrices}

\item{clust.data.test}{The test list of matrices}
}
\description{
Title
}
