\name{MMP-package}
\alias{MMP-package}
\alias{MMP}
\docType{package}
\title{
\packageTitle{MMP}
}
\description{
\packageDescription{MMP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MMP}
\packageIndices{MMP}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{MMP}

Maintainer: \packageMaintainer{MMP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
