% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintModels.r
\name{mmp}
\alias{mmp}
\title{Multi-model printing}
\usage{
mmp(AllModels, CoefNames = "<Generic>", Decimals = 4,
  WriteToClipboard = F, CoefOnly = F, CoefTests = NULL,
  Factors = NULL, BottomMatter = c("r.squared", "Brier", "N"),
  ModelNames = NULL, ShowDependentVariable = F, ShowModelType = F)
}
\arguments{
\item{AllModels}{List containing all model output objects. If the objects are named, the names will be used as headers.}

\item{CoefNames}{List with coefficient names.}

\item{ShowModelType}{}
}
\description{
Multi-model printing
}
