% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.CoefTest}
\alias{ModelPrintObject.CoefTest}
\title{ModelPrintObject.CoefTest}
\usage{
\method{ModelPrintObject}{CoefTest}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.CoefTest
}
