% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.data.frame}
\alias{ModelPrintObject.data.frame}
\title{ModelPrintObject.data.frame}
\usage{
\method{ModelPrintObject}{data.frame}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.data.frame
}
