% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.default}
\alias{ModelPrintObject.default}
\title{ModelPrintObject.default}
\usage{
\method{ModelPrintObject}{default}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.default
}
