% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.DiscreteEffects}
\alias{ModelPrintObject.DiscreteEffects}
\title{ModelPrintObject.DiscreteEffects}
\usage{
\method{ModelPrintObject}{DiscreteEffects}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.DiscreteEffects
}
