% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.fixest}
\alias{ModelPrintObject.fixest}
\title{ModelPrintObject.fixest}
\usage{
\method{ModelPrintObject}{fixest}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.fixest
}
