% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.glmerMod}
\alias{ModelPrintObject.glmerMod}
\title{ModelPrintObject.glmerMod}
\usage{
\method{ModelPrintObject}{glmerMod}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.glmerMod
}
