% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.lmerMod}
\alias{ModelPrintObject.lmerMod}
\title{ModelPrintObject.lmerMod}
\usage{
\method{ModelPrintObject}{lmerMod}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.lmerMod
}
