% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.lrm}
\alias{ModelPrintObject.lrm}
\title{ModelPrintObject.lrm}
\usage{
\method{ModelPrintObject}{lrm}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.lrm
}
