% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.ols}
\alias{ModelPrintObject.ols}
\title{ModelPrintObject.ols}
\usage{
\method{ModelPrintObject}{ols}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.ols
}
