% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.plm}
\alias{ModelPrintObject.plm}
\title{ModelPrintObject.plm}
\usage{
\method{ModelPrintObject}{plm}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.plm
}
