% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject}
\alias{ModelPrintObject}
\title{Extract model information from different classes}
\usage{
ModelPrintObject(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Extract model information from different classes
}
