% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.rdrobust}
\alias{ModelPrintObject.rdrobust}
\title{ModelPrintObject.rdrobust}
\usage{
\method{ModelPrintObject}{rdrobust}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.rdrobust
}
