% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.rms}
\alias{ModelPrintObject.rms}
\title{ModelPrintObject.rms}
\usage{
\method{ModelPrintObject}{rms}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.rms
}
