% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.selection}
\alias{ModelPrintObject.selection}
\title{ModelPrintObject.selection}
\usage{
\method{ModelPrintObject}{selection}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.selection
}
