% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.summary.felm}
\alias{ModelPrintObject.summary.felm}
\title{ModelPrintObject.felm}
\usage{
\method{ModelPrintObject}{summary.felm}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.felm
}
