% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.SummaryStat}
\alias{ModelPrintObject.SummaryStat}
\title{ModelPrintObject.SummaryStat}
\usage{
\method{ModelPrintObject}{SummaryStat}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.SummaryStat
}
