% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrintObjects.R
\name{ModelPrintObject.tobit}
\alias{ModelPrintObject.tobit}
\title{ModelPrintObject.boot}
\usage{
\method{ModelPrintObject}{tobit}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
ModelPrintObject.boot
}
