% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_func_file.R
\name{getDatalist}
\alias{getDatalist}
\title{Convert the file to the datalist}
\usage{
getDatalist(flist, idx.sel, ex.col = NULL)
}
\arguments{
\item{flist}{a vector containing the names of the feature files}

\item{idx.sel}{the indices to indicate the training data}

\item{ex.col}{the indices to exclude the columns}
}
\value{
a data list
}
\description{
Convert the file to the datalist
}
